:mouse: feat: Add `MouseEventImpl` to deal with canvas manipulation
Change `ICurvePointLocation` type to fit usecase, it doesn't extends `IPointBidimensional` anymore

+++ odontoradiosis_core/lib/src/controllers/subcontrollers/tracing_controller.dart
@@ -240,7 +240,7 @@ class TracingController extends AbstractBezierController {
             relativeMouse.y <= element[subIndex + 1] + pointRadius) {
           return ICurvePointLocation(
             element: element,
-            x: subIndex.toDouble(),
+            x: subIndex,
             y: subIndex + 1,
           );
         }
